~Moved register/scope to own function
